Fixed conversion bug

+++ Webservice/Models/Anchors.cs
@@ -6,12 +6,12 @@ using Microsoft.EntityFrameworkCore.SqlServer;
 
 namespace webservice.Model
 {
-    public class Anchor
+    public class Anchor 
     {
-            public long Id { get; set; }
+            public int Id { get; set; }
             public string Demo { get; set; }
             public string AnchorKey { get; set; }
-            public long AnchorType { get; set; }
+            public int AnchorType { get; set; }
             public DateTime Timestamp { get; set; }
     }
 }
